add iterAll fixed return float size dot two

+++ internal/repository/models.go
@@ -51,8 +51,8 @@ type Balance struct {
 }
 
 type ResponseBalance struct {
-	Current   float32 `json:"current"`
-	Withdrawn float32 `json:"withdrawn"`
+	Current   float64 `json:"current"`
+	Withdrawn float64 `json:"withdrawn"`
 }
 
 type CatalogData struct {
